moved ui to network side moved update check to global location
git-svn-id: https://aux1.edublogs.org/repos/incsub/trunk/plugins/autoblog@2028 f53fe46e-8c02-459b-9fdd-93e75771b278

+++ autoblogincludes/classes/autoblogadmin.php
@@ -29,7 +29,13 @@ class autoblogpremium {
 		add_action('load-autoblog_page_autoblog_admin', array(&$this, 'add_admin_header_autoblog_admin'));
 		add_action('load-autoblog_page_autoblog_options', array(&$this, 'add_admin_header_autoblog_options'));
 
-		add_action('admin_menu', array(&$this,'add_adminmenu'));
+		if(function_exists('is_multisite') && is_multisite()) {
+			if(function_exists('is_network_admin')) {
+				add_action('network_admin_menu', array(&$this,'add_adminmenu'));
+			} else {
+				add_action('admin_menu', array(&$this,'add_adminmenu'));
+			}
+		}
 
 		foreach($this->tables as $table) {
 			$this->$table = autoblog_db_prefix($this->db, $table);
@@ -93,19 +99,8 @@ class autoblogpremium {
 
 	}
 
-	function wdp_un_check() {
-	    if ( !class_exists( 'WPMUDEV_Update_Notifications' ) && current_user_can( 'edit_users' ) )
-	      echo '<div class="error fade"><p>' . __('Please install the latest version of <a href="http://premium.wpmudev.org/project/update-notifications/" title="Download Now &raquo;">our free Update Notifications plugin</a> which helps you stay up-to-date with the most stable, secure versions of WPMU DEV themes and plugins. <a href="http://premium.wpmudev.org/wpmu-dev/update-notifications-plugin-information/">More information &raquo;</a>', 'wpmudev') . '</a></p></div>';
-	  }
-
 	function add_update_check() {
-		/* -------------------- Update Notifications Notice -------------------- */
-		if ( method_exists( $this, 'wdp_un_check' ) ) {
-		  add_action( 'admin_notices', array(&$this, 'wdp_un_check'), 5 );
-		  add_action( 'network_admin_notices', array(&$this, 'wdp_un_check'), 5 );
 
-		}
-		/* --------------------------------------------------------------------- */
 	}
 
 	function add_admin_header_autoblog() {
@@ -1605,7 +1600,12 @@ class autoblogpremium {
 					echo "<input type='checkbox' name='select[]' id='select-" . $table->feed_id . "' value='" . $table->feed_id . "' class='selectfeed' />";
 					echo '</td>';
 					echo '<td>';
-					echo '<a href="' . admin_url("ms-admin.php?page=autoblog_admin&amp;edit=" . $table->feed_id) . '">';
+					if(function_exists('is_network_admin') && is_network_admin() ) {
+						echo '<a href="' . network_admin_url("admin.php?page=autoblog_admin&amp;edit=" . $table->feed_id) . '">';
+					} else {
+						echo '<a href="' . admin_url("admin.php?page=autoblog_admin&amp;edit=" . $table->feed_id) . '">';
+					}
+
 					if(!empty($details)) {
 						echo esc_html(stripslashes($details['title']));
 					} else {
@@ -1613,10 +1613,17 @@ class autoblogpremium {
 					}
 					echo '</a>';
 
+					//network_admin_url
 					echo '<div class="row-actions">';
-					echo "<a href='" . admin_url("admin.php?page=autoblog_admin&amp;edit=" . $table->feed_id) . "' class='editfeed'>" . __('Edit', 'autoblogtext') . "</a> | ";
-					echo "<a href='" . wp_nonce_url(admin_url("admin.php?page=autoblog_admin&amp;delete=" . $table->feed_id), 'autoblogdelete') . "' class='deletefeed'>" . __('Delete', 'autoblogtext') . "</a> | ";
-					echo "<a href='" . wp_nonce_url(admin_url("admin.php?page=autoblog_admin&amp;process=" . $table->feed_id), 'autoblogprocess') . "' class='processfeed'>" . __('Process', 'autoblogtext') . "</a>";
+					if(function_exists('is_network_admin') && is_network_admin() ) {
+						echo "<a href='" . network_admin_url("admin.php?page=autoblog_admin&amp;edit=" . $table->feed_id) . "' class='editfeed'>" . __('Edit', 'autoblogtext') . "</a> | ";
+						echo "<a href='" . wp_nonce_url(network_admin_url("admin.php?page=autoblog_admin&amp;delete=" . $table->feed_id), 'autoblogdelete') . "' class='deletefeed'>" . __('Delete', 'autoblogtext') . "</a> | ";
+						echo "<a href='" . wp_nonce_url(network_admin_url("admin.php?page=autoblog_admin&amp;process=" . $table->feed_id), 'autoblogprocess') . "' class='processfeed'>" . __('Process', 'autoblogtext') . "</a>";
+						} else {
+						echo "<a href='" . admin_url("admin.php?page=autoblog_admin&amp;edit=" . $table->feed_id) . "' class='editfeed'>" . __('Edit', 'autoblogtext') . "</a> | ";
+						echo "<a href='" . wp_nonce_url(admin_url("admin.php?page=autoblog_admin&amp;delete=" . $table->feed_id), 'autoblogdelete') . "' class='deletefeed'>" . __('Delete', 'autoblogtext') . "</a> | ";
+						echo "<a href='" . wp_nonce_url(admin_url("admin.php?page=autoblog_admin&amp;process=" . $table->feed_id), 'autoblogprocess') . "' class='processfeed'>" . __('Process', 'autoblogtext') . "</a>";
+					}
 					echo '</div>';
 
 					echo '</td>';
